requirements.txt
matplotlib==2.1.0
setuptools==39.2.0
requests==2.21.0
SQLAlchemy==1.2.13
bottle==0.12.13
tqdm==4.19.5
xmltodict==0.11.0
numpy==1.15.4
simplejson==3.8.1
Pillow==5.4.1
